Rename socketIo Js files
